requirements.txt
pymupdf>=1.22.0
opencv-python>=4.5.0
numpy>=1.19.0
paddleocr>=2.6.0
streamlit>=1.32.0
Pillow>=10.0.0
